# packages/kor-travel-map-admin/openapi.user.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PublicBeachDetailResponse": {
        "additionalProperties": false,
        "description": "``GET /public/beaches/{feature_id}`` 응답.",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/BeachPublicView"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "PublicBeachDetailResponse",
        "type": "object"
      },
      "PublicBeachListData": {
        "additionalProperties": false,
        "description": "해수욕장 공개 목록 data payload.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BeachPublicView"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "PublicBeachListData",
        "type": "object"
      },
      "PublicBeachListResponse": {
        "additionalProperties": false,
        "description": "``GET /public/beaches`` 응답.",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PublicBeachListData"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "PublicBeachListResponse",
        "type": "object"
      },
...
      "PublicMapMarkerLayerResponse": {
        "additionalProperties": false,
        "description": "공개 지도 layer marker 응답.",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PublicMapMarkerLayerData"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "PublicMapMarkerLayerResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "ServiceToken": {
        "description": "TripMate 등 외부 서비스 호출 토큰 (ADR-045 D-1).",
        "in": "header",
        "name": "X-Kor-Travel-Map-Service-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "TripMate/user-facing subset of kor-travel-map OpenAPI. Internal admin/debug/ops routes are intentionally excluded.",
    "title": "kor-travel-map-user",
    "version": "0.2.0-dev"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/public/beaches": {
      "get": {
        "description": "해수욕장 공개 목록 view.",
        "operationId": "list_public_beaches_v1_public_beaches_get",
        "parameters": [
          {
            "in": "query",
            "name": "sido_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 2,
                  "minLength": 2,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sido Code"
            }
          },
          {
            "in": "query",
            "name": "sigungu_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 5,
                  "minLength": 5,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sigungu Code"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "include_quality",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Quality",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "include_forecast",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Forecast",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicBeachListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Public Beaches",
        "tags": [
          "public"
        ]
      }
    },
    "/v1/public/beaches/map-markers": {
      "get": {
        "description": "해수욕장 공개 지도 marker layer.",
        "operationId": "list_public_beach_markers_v1_public_beaches_map_markers_get",
        "parameters": [
          {
            "in": "query",
            "name": "min_lon",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 180,
                  "minimum": -180,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Lon"
            }
          },
          {
            "in": "query",
            "name": "min_lat",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 90,
                  "minimum": -90,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Lat"
            }
          },
          {
            "in": "query",
            "name": "max_lon",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 180,
                  "minimum": -180,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Lon"
            }
          },
          {
            "in": "query",
            "name": "max_lat",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 90,
                  "minimum": -90,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Lat"
            }
          },
          {
            "in": "query",
            "name": "sido_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 2,
                  "minLength": 2,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sido Code"
            }
          },
          {
            "in": "query",
            "name": "sigungu_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 5,
                  "minLength": 5,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sigungu Code"
            }
          },
          {
            "in": "query",
            "name": "max_items",
            "required": false,
            "schema": {
              "default": 500,
              "maximum": 2000,
              "minimum": 1,
              "title": "Max Items",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicMapMarkerLayerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Public Beach Markers",
        "tags": [
          "public"
        ]
      }
    },
    "/v1/public/beaches/{feature_id}": {
      "get": {
        "description": "해수욕장 공개 상세 view.",
        "operationId": "get_public_beach_v1_public_beaches__feature_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "feature_id",
            "required": true,
            "schema": {
              "title": "Feature Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_quality",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Quality",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "include_forecast",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Forecast",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicBeachDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Public Beach",
        "tags": [
          "public"
        ]
      }
    },
    "/v1/public/festivals/map-markers": {
      "get": {
        "description": "축제 공개 지도 marker layer.",
        "operationId": "list_public_festival_markers_v1_public_festivals_map_markers_get",
        "parameters": [
          {
            "in": "query",
            "name": "year",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 2200,
                  "minimum": 1900,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          },
          {
            "in": "query",
            "name": "month",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 12,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          },
          {
            "in": "query",
            "name": "min_lon",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 180,
                  "minimum": -180,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Lon"
            }
          },
          {
            "in": "query",
            "name": "min_lat",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 90,
                  "minimum": -90,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Lat"
            }
          },
          {
            "in": "query",
            "name": "max_lon",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 180,
                  "minimum": -180,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Lon"
            }
          },
          {
            "in": "query",
            "name": "max_lat",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 90,
                  "minimum": -90,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Lat"
            }
          },
          {
            "in": "query",
            "name": "max_items",
            "required": false,
            "schema": {
              "default": 500,
              "maximum": 2000,
              "minimum": 1,
              "title": "Max Items",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicMapMarkerLayerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Public Festival Markers",
        "tags": [
          "public"
        ]
      }
    },
...
  }
}